create custom useExerciseNotification hook
