+++ src/components/sidebar/MainNav/SideBar/SidebarOptions_Student.js
@@ -1,9 +1,5 @@
-import { useState, useEffect, useContext } from "react";
 import { useLocation } from "react-router-dom/cjs/react-router-dom";
-import { APIContext } from "../../../../contexts/APIContext";
-import { ExerciseCountContext } from "../../../../exercises/ExerciseCountContext";
-import { userHasNotExercisedToday } from "../../../../exercises/utils/daysSinceLastExercise";
-import { MAX_EXERCISE_TO_DO_NOTIFICATION } from "../../../../exercises/ExerciseConstants";
+import useExerciseNotification from "../../../../hooks/useExerciseNotification";
 import NotificationIcon from "../../../NotificationIcon";
 import NavOption from "../NavOption";
 import HomeRoundedIcon from "@mui/icons-material/HomeRounded";
@@ -18,29 +14,11 @@ export default function SidebarOptions_Student({
   currentPath,
   isTeacher,
 }) {
-  const api = useContext(APIContext);
-  const exerciseNotification = useContext(ExerciseCountContext);
-  const [hasExerciseNotification, setHasExerciseNotification] = useState(false);
-  const [totalExercisesInPipeline, setTotalExercisesInPipeline] = useState();
   const path = useLocation().pathname;
 
-  useEffect(() => {
-    if (userHasNotExercisedToday() && path !== "/exercises")
-      api.getUserBookmarksToStudy(1, (scheduledBookmaks) => {
-        exerciseNotification.setHasExercises(scheduledBookmaks.length > 0);
-        exerciseNotification.updateReactState();
-      });
-    else {
-      exerciseNotification.setHasExercises(false);
-      exerciseNotification.updateReactState();
-    }
-  }, [path]);
+  const { hasExerciseNotification, notificationMsg } =
+    useExerciseNotification();
 
-  useEffect(() => {
-    exerciseNotification.setHasExercisesHook = setHasExerciseNotification;
-    exerciseNotification.setExerciseCounterHook = setTotalExercisesInPipeline;
-    exerciseNotification.updateReactState();
-  }, []);
   return (
     <>
       <NavOption
@@ -58,17 +36,7 @@ export default function SidebarOptions_Student({
         text={"Exercises"}
         currentPath={currentPath}
         notification={
-          hasExerciseNotification && (
-            <NotificationIcon
-              text={
-                totalExercisesInPipeline
-                  ? totalExercisesInPipeline > MAX_EXERCISE_TO_DO_NOTIFICATION
-                    ? MAX_EXERCISE_TO_DO_NOTIFICATION + "+"
-                    : totalExercisesInPipeline
-                  : ""
-              }
-            />
-          )
+          hasExerciseNotification && <NotificationIcon text={notificationMsg} />
         }
       />
